Merge branch 'feature/timeline-ui' into feature/timeline-ui-road-context-ui

# Conflicts:
#	DriverDataTimelineUI/src/main/java/com/drivequant/drivekit/timeline/ui/timeline/TimelineFragment.kt
#	DriverDataTimelineUI/src/main/java/com/drivequant/drivekit/timeline/ui/timeline/TimelineViewModel.kt

+++ DriverDataTimelineUI/src/main/java/com/drivequant/drivekit/timeline/ui/component/roadcontext/CustomRoadContextBar.kt
@@ -77,5 +77,5 @@ internal class CustomRoadContextBar @JvmOverloads constructor(
 
 internal data class ProgressItem(
     val color: Int,
-    val progressItemPercentage: Float
+    val progressItemPercentage: Double
 )
